Add ResultSetFactory.
The code for keys nesting and entities hydration has been extracted
from ResultSet into ResultSetFactory.
